requirements_info.txt
# Core Web Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
pydantic==2.6.1
jinja2==3.1.3
websockets==12.0
aiohttp==3.9.3
aiofiles==23.2.1

# Document Processing
python-docx==1.1.0
pandas==2.2.0
pdfminer.six==20231228
python-magic==0.4.27
PyMuPDF==1.23.8

# NLP and Analysis
spacy==3.7.2
en_core_web_sm>=3.0.0
textblob==0.17.1
gensim==4.3.2
langdetect==1.0.9
sentence-transformers==2.5.0

# Install spaCy model separately
# python -m spacy download en_core_web_sm

# ML Dependencies optimized for M3
--find-links https://download.pytorch.org/whl/nightly/cpu
torch==2.2.0
transformers==4.37.2
numpy==1.26.3
scipy==1.12.0
scikit-learn==1.4.0

# Visualization
plotly==5.18.0
wordcloud==1.9.3

# System and Performance
psutil==5.9.8
memory-profiler==0.61.0
tenacity==8.2.3
prometheus-client==0.19.0
cachetools==5.3.2

# Development and Testing
pytest==8.0.0
pytest-asyncio==0.23.5
pytest-cov==4.1.0
black==24.1.1

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1